Fix through-relation eager loading and update database details
- HasOneOrManyThrough read the far parent's key before checking whether
  constraints apply. While eager loading, the far parent is an empty
  model, so a strict key cast or accessor threw
  (https://github.com/laravel/framework/pull/59231). The key is now read
  only when constraints apply. A regression test eager loads through a
  model whose key accessor rejects a missing value.
- Morph class PHPDoc on MorphOneOrMany, MorphPivot and MorphToMany
  accepts morph-map aliases (https://github.com/laravel/framework/pull/59110).
- Lost-connection detection labels the PlanetScale MySQL messages and
  adds the PlanetScale PostgreSQL / pg_bouncer messages
  (https://github.com/laravel/framework/pull/59244 and Laravel commit
  7fffd3e917).
- The Testbench cache and cache_locks migrations store expiration as an
  unsigned big integer, matching Hypervel's skeleton and generated
  migrations (https://github.com/laravel/framework/pull/59243). A 32-bit
  column overflows once forever()'s ten-year expiry passes 2038.
- Mark the removed SQL Server unique-constraint tests
  (https://github.com/laravel/framework/pull/59299).

Validation: the through-relation tests pass, and the new regression
fails without the fix. The database cache and lock tests pass on SQLite,
MySQL and PostgreSQL, as does the Testbench suite.

+++ src/testbench/hypervel/migrations/0001_01_01_000003_testbench_create_cache_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
         Schema::create('cache', function (Blueprint $table) {
             $table->string('key')->primary();
             $table->mediumText('value');
-            $table->integer('expiration')->index();
+            $table->unsignedBigInteger('expiration')->index();
         });
     }
